@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class GraphqlController < ApplicationController
+class GraphQLController < ApplicationController
   skip_before_action :verify_authenticity_token
 
   include AuthenticateUser
