@@ -1,5 +1,7 @@
 module Mutations
   class User::DisconnectSlack < BaseMutation
+    null true
+
     field :user, Types::UserType, null: true
 
     def resolve()
@@ -9,7 +11,7 @@ module Mutations
       user.slack_id = nil
 
       if user.save
-        {user: user}
+        { user: user }
       else
         return Util::ErrorBuilder.build_errors(context, user.errors)
       end
