@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 module Types
-  class BaseInputObject < GraphQL::Schema::InputObject; end
+  class BaseInputObject < GraphQL::Schema::InputObject
+  end
 end
