@@ -1,19 +1,20 @@
 # frozen_string_literal: true
 
 module Types
-  Date = GraphQL::ScalarType.define do
-    name "Date"
+  class Date < Types::BaseScalar
+    graphql_name "Date"
     description "An ISO 8601-encoded date"
 
-    coerce_input ->(value, ctx) do
+    def self.coerce_input(value, _context)
       begin
-        Date.iso8601(value.to_s)
+        ::Date.iso8601(value.to_s)
       rescue ArgumentError
         raise GraphQL::CoercionError, "cannot coerce `#{value.inspect}` to Date"
       end
     end
 
-    coerce_result ->(value, ctx) { value }
+    def self.coerce_result(value, _context)
+      value
+    end
   end
 end
-
