@@ -14,7 +14,7 @@ module Types
     field :achieved_on, Types::Date,
           null: true,
           description: 'The date the goal is achieved'
-    field :kudosMeter, Types::KudosMeterType,
+    field :kudos_meter, Types::KudosMeterType,
           null: false,
           description: 'The kudos meter the goal belongs to',
           resolve: ->(obj, _args, _ctx) { Util::RecordLoader.for(KudosMeter).load(obj.kudos_meter_id) }
